# contracts/openapi/api-v1.json
{
  "components": {
    "schemas": {
      "BootstrapResponse": {
        "properties": {
          "navigation": {
            "items": {
              "$ref": "#/components/schemas/NavigationNode"
            },
            "type": "array"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "navigation",
          "permissions",
          "roles",
          "user"
        ],
        "type": "object"
      },
      "DeletePostsRequest": {
        "properties": {
          "ids": {
            "items": {
              "pattern": "[1-9][0-9]{0,18}",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "code": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "uuid": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "tokenType"
        ],
        "type": "object"
      },
      "NavigationNode": {
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/NavigationNode"
            },
            "type": "array"
          },
          "externalUrl": {
            "description": "Present only for EXTERNAL nodes",
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "order": {
            "format": "int32",
            "type": "integer"
          },
          "routeId": {
            "description": "Present only for ROUTE nodes",
            "type": "string"
          },
          "type": {
            "enum": [
              "GROUP",
              "ROUTE",
              "EXTERNAL"
            ],
            "type": "string"
          }
        },
        "required": [
          "children",
          "key",
          "label",
          "order",
          "type"
        ],
        "type": "object"
      },
      "PageResponsePostResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object"
      },
      "PostRequest": {
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "remark": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "sort": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "pattern": "[01]",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "sort",
          "status"
        ],
        "type": "object"
      },
      "PostResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "sort": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "id",
          "name",
          "sort",
          "status"
        ],
        "type": "object"
      },
      "ProblemDetail": {
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "format": "uri",
            "type": [
              "string",
              "null"
            ]
          },
          "properties": {
            "additionalProperties": {},
            "type": [
              "object",
              "null"
            ]
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "UserSummary": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "id",
          "username"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "apikey": {
        "in": "header",
        "name": "Authorization",
        "scheme": "Bearer",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "EForge Enterprise"
    },
    "description": "描述：用于管理集团旗下公司的人员信息,具体包括XXX,XXX模块...",
    "title": "标题：若依管理系统_接口文档",
    "version": "版本号:0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/app/bootstrap": {
      "get": {
        "operationId": "bootstrap",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the current user and authorized application navigation",
        "tags": [
          "bootstrap-controller"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Authenticated"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Invalid request or captcha"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Login rejected"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Login blocked"
          }
        },
        "security": [],
        "summary": "Create a Redis-backed login session",
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/v1/system/posts": {
      "delete": {
        "operationId": "deletePosts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletePostsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Posts deleted"
          }
        },
        "tags": [
          "post-controller"
        ]
      },
      "get": {
        "operationId": "listPosts",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "maximum": 1000000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 64,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "[01]?",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePostResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "post-controller"
        ]
      },
      "post": {
        "operationId": "createPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            },
            "description": "Created post"
          }
        },
        "tags": [
          "post-controller"
        ]
      }
    },
    "/api/v1/system/posts/export": {
      "post": {
        "operationId": "exportPosts",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 64,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "[01]?",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Filtered XLSX workbook"
          }
        },
        "tags": [
          "post-controller"
        ]
      }
    },
    "/api/v1/system/posts/{id}": {
      "get": {
        "operationId": "getPost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "[1-9][0-9]{0,18}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "post-controller"
        ]
      },
      "put": {
        "operationId": "updatePost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "[1-9][0-9]{0,18}",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "post-controller"
        ]
      }
    }
  },
  "security": [
    {
      "apikey": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ]
}